refactored cadc-stiorage-adapter-fs and -ceph intTest code into common test lib
