+++ cadc-storage-adapter-ceph/src/intTest/java/org/opencadc/inventory/storage/swift/SwiftByteRangeTest.java
@@ -92,6 +92,7 @@ import org.opencadc.inventory.StorageLocation;
 import org.opencadc.inventory.storage.ByteRange;
 import org.opencadc.inventory.storage.NewArtifact;
 import org.opencadc.inventory.storage.StorageMetadata;
+import org.opencadc.inventory.storage.test.StorageAdapterByteRangeTest;
 
 /**
  * Integration tests that interact with the file system. These tests require a file system
@@ -99,155 +100,30 @@ import org.opencadc.inventory.storage.StorageMetadata;
  * 
  * @author pdowler
  */
-public class SwiftByteRangeTest {
+public class SwiftByteRangeTest extends StorageAdapterByteRangeTest {
     private static final Logger log = Logger.getLogger(SwiftByteRangeTest.class);
 
     static {
         Log4jInit.setLevel("org.opencadc.inventory", Level.INFO);
-        Log4jInit.setLevel("org.javaswift.joss", Level.INFO);
+        //Log4jInit.setLevel("org.javaswift.joss", Level.INFO);
     }
     
-    final SwiftStorageAdapter adapter;
+    final SwiftStorageAdapter swiftAdapter;
     
     public SwiftByteRangeTest() {
-        this.adapter = new SwiftStorageAdapter();
+        super(new SwiftStorageAdapter());
+        this.swiftAdapter = (SwiftStorageAdapter) super.adapter;
     }
     
     @Before
-    public void cleanup() throws Exception {
-        final long t1 = System.currentTimeMillis();
-        log.info("cleanup: ");
-        Iterator<StorageMetadata> sbi = adapter.iterator();
+    public void cleanupBefore() throws Exception {
+        log.info("cleanupBefore: START");
+        Iterator<StorageMetadata> sbi = swiftAdapter.iterator();
         while (sbi.hasNext()) {
             StorageLocation loc = sbi.next().getStorageLocation();
-            adapter.delete(loc);
+            swiftAdapter.delete(loc);
             log.info("\tdeleted: " + loc);
         }
-    }
-    
-    @Test
-    public void testFullRead() {
-        try {
-            int mib = 16;
-            long datalen = mib * 1024L * 1024L; // 1 MiB
-            
-            URI artifactURI = URI.create("test:path/file");
-            NewArtifact newArtifact = new NewArtifact(artifactURI);
-            InputStream istream = TestUtil.getInputStreamOfRandomBytes(datalen);
-
-            log.info("put: " + mib + " MiB file...");
-            long t1 = System.nanoTime();
-            StorageMetadata storageMetadata = adapter.put(newArtifact, istream);
-            long t2 = System.nanoTime();
-            final long putMicros = (t2 - t1) / 1024L;
-            StringBuilder sb = new StringBuilder();
-            sb.append("put: ").append(storageMetadata.getStorageLocation()).append(" -- ");
-            
-            sb.append(Long.toString(putMicros)).append(" microsec");
-            double spd = (double) (10 * datalen / putMicros) / 10.0;
-            sb.append(" aka ~").append(spd).append(" MiB/sec");
-            log.info(sb);
-            
-            Assert.assertNotNull(storageMetadata);
-            Assert.assertNotNull(storageMetadata.getStorageLocation());
-            Assert.assertEquals(datalen, storageMetadata.getContentLength().longValue());
-            
-            
-            ByteRange r = new ByteRange(0, datalen);
-
-            SortedSet<ByteRange> br = new TreeSet<>();
-            br.add(r);
-            DigestOutputStream out = new DigestOutputStream(new DiscardOutputStream(), MessageDigest.getInstance("MD5"));
-            ByteCountOutputStream bcos = new ByteCountOutputStream(out);
-            t1 = System.nanoTime();
-            adapter.get(storageMetadata.getStorageLocation(), bcos);
-            //adapter.get(storageMetadata.getStorageLocation(), bcos, br);
-            t2 = System.nanoTime();
-            final long getMicros = (t2 - t1) / 1024L;
-            Assert.assertEquals("num bytes returned", datalen, bcos.getByteCount());
-            URI actualMD5 = URI.create("md5:" + HexUtil.toHex(out.getMessageDigest().digest()));
-            Assert.assertEquals("checksum", storageMetadata.getContentChecksum(), actualMD5);
-            sb = new StringBuilder();
-            sb.append("read ").append(r);
-            while (sb.length() < 32) {
-                sb.append(" ");
-            }
-            sb.append(Long.toString(getMicros)).append(" microsec");
-            spd = (double) (10 * datalen / getMicros) / 10.0;
-            sb.append(" aka ~").append(spd).append(" MiB/sec");
-            log.info(sb);
-            
-            adapter.delete(storageMetadata.getStorageLocation());
-            
-        } catch (Exception unexpected) {
-            log.error("unexpected exception", unexpected);
-            Assert.fail("unexpected exception: " + unexpected);
-        }
-    }
-    
-    @Test
-    public void testReadByteRangeForward() {
-        int[] readOrder = new int[] {0, 1, 2, 0, 1, 2, 0, 1, 2};
-        doReadPattern(readOrder);
-    }
-    
-    @Test
-    public void testReadByteRangeReverse() {
-        int[] readOrder = new int[] {2, 1, 0, 2, 1, 0, 2, 1, 0};
-        doReadPattern(readOrder);
-    }
-    
-    public void doReadPattern(int[] readOrder) {
-        try {
-            int mib = 16;
-            long datalen = mib * 1024L * 1024L; // 1 MiB
-            
-            URI artifactURI = URI.create("test:path/file");
-            NewArtifact newArtifact = new NewArtifact(artifactURI);
-            InputStream istream = TestUtil.getInputStreamOfRandomBytes(datalen);
-
-            log.info("put: " + mib + " MiB file...");
-            StorageMetadata storageMetadata = adapter.put(newArtifact, istream);
-            log.info("put: " + storageMetadata.getStorageLocation());
-            Assert.assertNotNull(storageMetadata);
-            Assert.assertNotNull(storageMetadata.getStorageLocation());
-            Assert.assertEquals(datalen, storageMetadata.getContentLength().longValue());
-            
-            log.info("sleep 5000 to let server settle down...");
-            Thread.sleep(5000L);
-            log.info("get: " + storageMetadata.getStorageLocation());
-            
-            List<ByteRange> ranges = new ArrayList<>();
-            long rlen = 16 * 1024L; // 16KiB
-            ranges.add(new ByteRange(0L, rlen));                    // at start
-            ranges.add(new ByteRange(datalen / 2L, rlen));          // near the middle
-            ranges.add(new ByteRange(datalen - rlen - 1L, rlen));   // at end
-            
-            for (int i : readOrder) {
-                ByteRange r = ranges.get(i);
-                
-                SortedSet<ByteRange> br = new TreeSet<>();
-                br.add(r);
-                ByteCountOutputStream bcos = new ByteCountOutputStream(new DiscardOutputStream());
-                long t1 = System.nanoTime();
-                adapter.get(storageMetadata.getStorageLocation(), bcos, br);
-                long t2 = System.nanoTime();
-                final long micros = (t2 - t1) / 1024L;
-                Assert.assertEquals("num bytes returned", rlen, bcos.getByteCount());
-                StringBuilder sb = new StringBuilder();
-                sb.append("read ").append(r);
-                while (sb.length() < 32) {
-                    sb.append(" ");
-                }
-                sb.append(Long.toString(micros)).append(" microsec");
-                log.info(sb);
-            }
-            
-            adapter.delete(storageMetadata.getStorageLocation());
-            
-        } catch (Exception unexpected) {
-            log.error("unexpected exception", unexpected);
-            Assert.fail("unexpected exception: " + unexpected);
-        }
+        log.info("cleanupBefore: DONE");        
     }
 }
