+++ cadc-storage-adapter-ceph/src/intTest/java/org/opencadc/inventory/storage/swift/SwiftStorageAdapterTest.java
@@ -68,26 +68,10 @@
 package org.opencadc.inventory.storage.swift;
 
 import ca.nrc.cadc.io.ByteLimitExceededException;
-import ca.nrc.cadc.io.ReadException;
-import ca.nrc.cadc.io.WriteException;
-import ca.nrc.cadc.net.PreconditionFailedException;
-import ca.nrc.cadc.net.ResourceNotFoundException;
-import ca.nrc.cadc.util.HexUtil;
 import ca.nrc.cadc.util.Log4jInit;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
 import java.io.InputStream;
 import java.net.URI;
-import java.security.MessageDigest;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Iterator;
-import java.util.List;
-import java.util.Random;
-import java.util.Set;
-import java.util.SortedSet;
-import java.util.TreeSet;
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
 import org.junit.Assert;
@@ -96,12 +80,14 @@ import org.junit.Test;
 import org.opencadc.inventory.StorageLocation;
 import org.opencadc.inventory.storage.NewArtifact;
 import org.opencadc.inventory.storage.StorageMetadata;
+import org.opencadc.inventory.storage.test.StorageAdapterBasicTest;
+import org.opencadc.inventory.storage.test.TestUtil;
 
 /**
  *
  * @author pdowler
  */
-public class SwiftStorageAdapterTest {
+public class SwiftStorageAdapterTest extends StorageAdapterBasicTest {
     private static final Logger log = Logger.getLogger(SwiftStorageAdapterTest.class);
 
     static {
@@ -109,206 +95,25 @@ public class SwiftStorageAdapterTest {
         Log4jInit.setLevel("org.javaswift.joss.client", Level.INFO);
     }
     
-    final SwiftStorageAdapter adapter;
+    final SwiftStorageAdapter swiftAdapter;
     
     public SwiftStorageAdapterTest() {
-        this.adapter = new SwiftStorageAdapter();
+        super(new SwiftStorageAdapter());
+        this.swiftAdapter = (SwiftStorageAdapter) super.adapter;
     }
     
     @Before
-    public void cleanup() throws Exception {
-        final long t1 = System.currentTimeMillis();
-        log.info("cleanup: ");
-        Iterator<StorageMetadata> sbi = adapter.iterator();
+    public void cleanupBefore() throws Exception {
+        log.info("cleanupBefore: START");
+        Iterator<StorageMetadata> sbi = swiftAdapter.iterator();
         while (sbi.hasNext()) {
             StorageLocation loc = sbi.next().getStorageLocation();
-            adapter.delete(loc);
+            swiftAdapter.delete(loc);
             log.info("\tdeleted: " + loc);
         }
+        log.info("cleanupBefore: DONE");        
     }
     
-    //@Test
-    public void testNoOp() {
-    }
-    
-    @Test
-    public void testResourceNotFound() {
-        try {
-            StorageLocation loc = adapter.generateStorageLocation();
-            adapter.get(loc, TestUtil.getOutputStreamThatFails());
-            Assert.fail("expected ResourceNotFoundException: call succeeded");
-        } catch (ResourceNotFoundException expected) {
-            log.info("caught expected: " + expected);
-        } catch (Exception unexpected) {
-            log.error("unexpected exception", unexpected);
-            Assert.fail("unexpected exception: " + unexpected);
-        }
-    }
-    
-    @Test
-    public void testPutReadFail() {
-        try {
-            URI artifactURI = URI.create("cadc:TEST/testPutReadFail");
-            NewArtifact na = new NewArtifact(artifactURI);
-            adapter.put(na, TestUtil.getInputStreamThatFails(true)); // provoke IOException
-            Assert.fail("expected ReadException: call succeeded");
-        } catch (ReadException expected) {
-            log.info("caught expected: " + expected);
-        } catch (Exception unexpected) {
-            log.error("unexpected exception", unexpected);
-            Assert.fail("unexpected exception: " + unexpected);
-        }
-    }
-    
-    @Test
-    public void testGetWriteFail() {
-        try {
-            URI artifactURI = URI.create("cadc:TEST/testPutReadFail");
-            NewArtifact na = new NewArtifact(artifactURI);
-            StorageMetadata sm = adapter.put(na, TestUtil.getInputStreamOfRandomBytes(8192));
-            Assert.assertNotNull(sm);
-            
-            adapter.get(sm.getStorageLocation(), TestUtil.getOutputStreamThatFails(true)); // provoke IOException
-            Assert.fail("expected WriteException: call succeeded");
-        } catch (WriteException expected) {
-            log.info("caught expected: " + expected);
-        } catch (Exception unexpected) {
-            log.error("unexpected exception", unexpected);
-            Assert.fail("unexpected exception: " + unexpected);
-        }
-    }
-    
-    @Test
-    public void testPutGetDelete() {
-        URI artifactURI = URI.create("cadc:TEST/testPutGetDelete");
-        
-        try {
-            MessageDigest md = MessageDigest.getInstance("MD5");
-            Random rnd = new Random();
-            byte[] data = new byte[1024];
-            rnd.nextBytes(data);
-            
-            NewArtifact na = new NewArtifact(artifactURI);
-            md.update(data);
-            URI expectedChecksum = URI.create("md5:" + HexUtil.toHex(md.digest()));
-            na.contentChecksum = expectedChecksum;
-            na.contentLength = (long) data.length;
-            log.debug("testPutGetDelete random data: " + data.length + " " + expectedChecksum);
-            
-            log.debug("testPutGetDelete put: " + artifactURI);
-            StorageMetadata sm = adapter.put(na, new ByteArrayInputStream(data));
-            log.info("testPutGetDelete put: " + artifactURI + " to " + sm.getStorageLocation());
-            
-            // verify metadata
-            StorageMetadata persisted = adapter.getMeta(sm.getStorageLocation());
-            Assert.assertNotNull(persisted);
-            Assert.assertTrue("valid", persisted.isValid());
-            Assert.assertEquals(na.contentChecksum, persisted.getContentChecksum());
-            Assert.assertEquals(na.contentLength, persisted.getContentLength());
-            Assert.assertEquals(na.getArtifactURI(), persisted.artifactURI);
-            
-            // verify data stored
-            log.debug("testPutGetDelete get: " + artifactURI);
-            ByteArrayOutputStream bos = new ByteArrayOutputStream();
-            adapter.get(sm.getStorageLocation(), bos);
-            md.reset();
-            byte[] actual = bos.toByteArray();
-            md.update(actual);
-            URI actualChecksum = URI.create("md5:" + HexUtil.toHex(md.digest()));
-            log.info("testPutGetDelete get: " + artifactURI + " " + actual.length + " " + actualChecksum);
-            Assert.assertEquals("length", (long) na.contentLength, actual.length);
-            Assert.assertEquals("checksum", na.contentChecksum, actualChecksum);
-
-            log.debug("testPutGetDelete delete: " + sm.getStorageLocation());
-            adapter.delete(sm.getStorageLocation());
-            Assert.assertTrue("deleted", !adapter.exists(sm.getStorageLocation()));
-            log.info("testPutGetDelete deleted: " + sm.getStorageLocation());
-        } catch (Exception ex) {
-            log.error("unexpected exception", ex);
-            Assert.fail("unexpected exception: " + ex);
-        }
-    }   
-    
-    @Test
-    public void testPutGetDeleteMinimal() {
-        URI artifactURI = URI.create("cadc:TEST/testPutGetDeleteMinimal");
-        
-        try {
-            Random rnd = new Random();
-            byte[] data = new byte[1024];
-            rnd.nextBytes(data);
-            
-            final NewArtifact na = new NewArtifact(artifactURI);
-            MessageDigest md = MessageDigest.getInstance("MD5");
-            md.update(data);
-            URI expectedChecksum = URI.create("md5:" + HexUtil.toHex(md.digest()));
-            
-            // test that we can store raw input stream data with no optional metadata
-            
-            log.debug("testPutGetDeleteMinimal random data: " + data.length + " " + expectedChecksum);
-            
-            log.debug("testPutGetDeleteMinimal put: " + artifactURI);
-            StorageMetadata sm = adapter.put(na, new ByteArrayInputStream(data));
-            log.info("testPutGetDeleteMinimal put: " + artifactURI + " to " + sm.getStorageLocation());
-            
-            // verify metadata
-            StorageMetadata persisted = adapter.getMeta(sm.getStorageLocation());
-            log.info("persisted: " + persisted);
-            Assert.assertNotNull(persisted);
-            Assert.assertTrue("valid", persisted.isValid());
-            Assert.assertEquals(expectedChecksum, persisted.getContentChecksum());
-            Assert.assertEquals(data.length, persisted.getContentLength().longValue());
-            Assert.assertEquals(na.getArtifactURI(), persisted.artifactURI);
-            
-            // verify data
-            log.debug("testPutGetDeleteMinimal get: " + artifactURI);
-            ByteArrayOutputStream bos = new ByteArrayOutputStream();
-            adapter.get(sm.getStorageLocation(), bos);
-            byte[] actual = bos.toByteArray();
-            md.update(actual);
-            URI actualChecksum = URI.create("md5:" + HexUtil.toHex(md.digest()));
-            log.info("testPutGetDeleteMinimal get: " + artifactURI + " " + actual.length + " " + actualChecksum);
-            Assert.assertEquals("length", data.length, actual.length);
-            Assert.assertEquals("checksum", expectedChecksum, actualChecksum);
-            
-            log.debug("testPutGetDeleteMinimal delete: " + sm.getStorageLocation());
-            adapter.delete(sm.getStorageLocation());
-            Assert.assertTrue("deleted", !adapter.exists(sm.getStorageLocation()));
-            log.info("testPutGetDeleteMinimal deleted: " + sm.getStorageLocation());
-        } catch (Exception ex) {
-            log.error("unexpected exception", ex);
-            Assert.fail("unexpected exception: " + ex);
-        }
-    }
-    
-    @Test
-    public void testPutGetDeleteWrongMD5() {
-        URI artifactURI = URI.create("cadc:TEST/testPutGetDeleteWrongMD5");
-        
-        try {
-            Random rnd = new Random();
-            byte[] data = new byte[1024];
-            rnd.nextBytes(data);
-            
-            NewArtifact na = new NewArtifact(artifactURI);
-            na.contentChecksum = URI.create("md5:d41d8cd98f00b204e9800998ecf8427e"); // md5 of 0-length file
-            na.contentLength = (long) data.length;
-            log.debug("testPutGetDeleteWrongMD5 random data: " + data.length + " " +  na.contentChecksum);
-            
-            try {
-                log.debug("testPutGetDeleteWrongMD5 put: " + artifactURI);
-                StorageMetadata sm = adapter.put(na, new ByteArrayInputStream(data));
-                Assert.fail("expected PreconditionFailedException, got: " + sm.getStorageLocation());
-            } catch (PreconditionFailedException expected) {
-                log.info("testPutGetDeleteWrongMD5 caught: " + expected);
-            }
-            
-        } catch (Exception ex) {
-            log.error("unexpected exception", ex);
-            Assert.fail("unexpected exception: " + ex);
-        }
-    }
-
     @Test
     public void testPutLargeStreamReject() {
         URI artifactURI = URI.create("cadc:TEST/testPutLargeStreamReject");
@@ -322,7 +127,7 @@ public class SwiftStorageAdapterTest {
         try {
             InputStream istream = TestUtil.getInputStreamThatFails();
             log.info("testPutCheckDeleteLargeStreamReject put: " + artifactURI + " " + numBytes);
-            StorageMetadata sm = adapter.put(na, istream);
+            StorageMetadata sm = swiftAdapter.put(na, istream);
             Assert.fail("expected ByteLimitExceededException, got: " + sm);
         } catch (ByteLimitExceededException expected) {
             log.info("caught: " + expected);
@@ -345,9 +150,9 @@ public class SwiftStorageAdapterTest {
         try {
             InputStream istream = TestUtil.getInputStreamOfRandomBytes(numBytes);
             log.info("testPutCheckDeleteLargeStreamFail put: " + artifactURI + " " + numBytes);
-            StorageMetadata sm = adapter.put(na, istream);
+            StorageMetadata sm = swiftAdapter.put(na, istream);
             
-            Assert.assertFalse("put should have failed, but object exists", adapter.exists(sm.getStorageLocation()));
+            Assert.assertFalse("put should have failed, but object exists", swiftAdapter.exists(sm.getStorageLocation()));
             
             Assert.fail("expected ByteLimitExceededException, got: " + sm);
         } catch (ByteLimitExceededException expected) {
@@ -358,90 +163,5 @@ public class SwiftStorageAdapterTest {
         }
     }
     
-    @Test
-    public void testIterator() {
-        int iterNum = 13;
-        long datalen = 8192L;
-        try {
-            SortedSet<StorageMetadata> expected = new TreeSet<>();
-            for (int i = 0; i < iterNum; i++) {
-                URI artifactURI = URI.create("cadc:TEST/testIterator-" + i);
-                NewArtifact na = new NewArtifact(artifactURI);
-                na.contentLength = (long) datalen;
-                StorageMetadata sm = adapter.put(na, TestUtil.getInputStreamOfRandomBytes(datalen));
-                Assert.assertNotNull(sm.artifactURI);
-                //Assert.assertNotNull(sm.contentLastModified);
-                log.debug("testIterator put: " + artifactURI + " to " + sm.getStorageLocation());
-                expected.add(sm);
-            }
-            log.info("testIterator created: " + expected.size());
-            
-            // full iterator
-            List<StorageMetadata> actual = new ArrayList<>();
-            Iterator<StorageMetadata> iter = adapter.iterator();
-            while (iter.hasNext()) {
-                StorageMetadata sm = iter.next();
-                log.debug("found: " + sm.getStorageLocation() + " " + sm.getContentLength() + " " + sm.getContentChecksum());
-                actual.add(sm);
-            }
-            
-            Assert.assertEquals("iterator.size", expected.size(), actual.size());
-            Iterator<StorageMetadata> ei = expected.iterator();
-            Iterator<StorageMetadata> ai = actual.iterator();
-            while (ei.hasNext()) {
-                StorageMetadata em = ei.next();
-                StorageMetadata am = ai.next();
-                log.debug("compare: " + em.getStorageLocation() + " vs " + am.getStorageLocation());
-                Assert.assertEquals("order", em, am);
-                Assert.assertEquals("length", em.getContentLength(), am.getContentLength());
-                Assert.assertEquals("checksum", em.getContentChecksum(), am.getContentChecksum());
-                
-                Assert.assertNotNull("artifactUIRI", am.artifactURI);
-                Assert.assertEquals("artifactURI", em.artifactURI, am.artifactURI);
-                
-                Assert.assertNotNull("contentLastModified", am.contentLastModified);
-                //Assert.assertEquals("contentLastModified", em.contentLastModified, am.contentLastModified);
-            }
-            
-            // rely on cleanup()
-        } catch (Exception ex) {
-            log.error("unexpected exception", ex);
-            Assert.fail("unexpected exception: " + ex);
-        }
-    }
     
-    @Test
-    public void testIteratorBucketPrefix() {
-        int iterNum = 13;
-        long datalen = 8192L;
-        try {
-            SortedSet<StorageMetadata> expected = new TreeSet<>();
-            for (int i = 0; i < iterNum; i++) {
-                URI artifactURI = URI.create("cadc:TEST/testIteratorBucketPrefix-" + i);
-                NewArtifact na = new NewArtifact(artifactURI);
-                na.contentLength = (long) datalen;
-                StorageMetadata sm = adapter.put(na,  TestUtil.getInputStreamOfRandomBytes(datalen));
-                log.debug("testList put: " + artifactURI + " to " + sm.getStorageLocation());
-                expected.add(sm);
-            }
-            log.info("testIteratorBucketPrefix created: " + expected.size());
-            
-            int found = 0;
-            for (byte b = 0; b < 16; b++) {
-                String bpre = HexUtil.toHex(b).substring(1);
-                log.debug("bucket prefix: " + bpre);
-                Iterator<StorageMetadata> i = adapter.iterator(bpre);
-                while (i.hasNext()) {
-                    StorageMetadata sm = i.next();
-                    Assert.assertTrue("prefix match", sm.getStorageLocation().storageBucket.startsWith(bpre));
-                    found++;
-                }
-            }
-            Assert.assertEquals("found with bucketPrefix", expected.size(), found);
-            
-        } catch (Exception ex) {
-            log.error("unexpected exception", ex);
-            Assert.fail("unexpected exception: " + ex);
-        }
-    }
 }
