+++ cadc-storage-adapter-test/README.md
@@ -1,0 +1,6 @@
+# cadc-storage-adapter-test
+
+Basic test suite for common StorageAdapter behaviour.
+
+StorageAdapter implementations subclass the abstract classes in this library ti add intTest(s).
+
